+++ daebbang-api/src/main/java/com/daebbang/daebbangapi/domain/user/dto/request/LoginRequest.java
@@ -1,0 +1,9 @@
+package com.daebbang.daebbangapi.domain.user.dto.request;
+
+public record LoginRequest(
+
+    String id,
+    String password
+) {
+
+}
