+++ daebbang-api/src/main/java/com/daebbang/daebbangapi/domain/user/dto/response/SmsSendAuthCode.java
@@ -1,0 +1,9 @@
+package com.daebbang.daebbangapi.domain.user.dto.response;
+
+public record SmsSendAuthCode(
+    String authCode
+) {
+    public static SmsSendAuthCode toDto(String authCode) {
+        return new SmsSendAuthCode(authCode);
+    }
+}
