+++ daebbang-api/src/main/java/com/daebbang/daebbangapi/domain/user/dto/response/UserIdFind.java
@@ -1,0 +1,20 @@
+package com.daebbang.daebbangapi.domain.user.dto.response;
+
+import com.daebbang.daebbangcore.domain.user.entity.Users;
+import java.util.List;
+import java.util.Objects;
+
+public record UserIdFind(
+    List<IdInfo> userIds
+) {
+    public static UserIdFind toDto(List<Users> users) {
+        if (Objects.isNull(users) || users.isEmpty()) {
+            return new UserIdFind(List.of());
+        }
+
+        List<IdInfo> maskedIdInfos = users.stream()
+                                        .map(info -> IdInfo.ofMasked(info.getLoginId(), info.getProvider()))
+                                        .toList();
+        return new UserIdFind(maskedIdInfos);
+    }
+}
