+++ daebbang-api/src/main/java/com/daebbang/daebbangapi/domain/user/service/CustomUserDetailsService.java
@@ -1,0 +1,31 @@
+package com.daebbang.daebbangapi.domain.user.service;
+
+import com.daebbang.daebbangcore.domain.user.entity.Users;
+import com.daebbang.daebbangcore.domain.user.service.UserService;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.jspecify.annotations.NullMarked;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@NullMarked
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final UserService userService;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Users user = userService.getUser(username);
+        return new User(
+            user.getLoginId(),
+            user.getLoginPwd(),
+            List.of(new SimpleGrantedAuthority("ROLE_USER"))
+        );
+    }
+}
