--- daebbang-api/src/main/java/com/daebbang/daebbangapi/domain/users/dto/response/SmsSendAuthCode.java
@@ -1,9 +1,0 @@
-package com.daebbang.daebbangapi.domain.users.dto.response;
-
-public record SmsSendAuthCode(
-    String authCode
-) {
-    public static SmsSendAuthCode toDto(String authCode) {
-        return new SmsSendAuthCode(authCode);
-    }
-}
