+++ daebbang-api/src/test/java/com/daebbang/daebbangapi/domain/review/controller/ReviewControllerTest.java
@@ -21,7 +21,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.daebbang.daebbangapi.config.PasswordConfig;
 import com.daebbang.daebbangapi.config.TestSecurityConfig;
 import com.daebbang.daebbangapi.domain.oauth.service.oauth2.Oauth2UserDetailsService;
-import com.daebbang.daebbangapi.domain.users.service.CustomUserDetailsService;
+import com.daebbang.daebbangapi.domain.user.service.CustomUserDetailsService;
 import com.daebbang.daebbangcore.domain.review.entity.Review;
 import com.daebbang.daebbangcore.domain.review.entity.ReviewPointConfig;
 import com.daebbang.daebbangcore.domain.review.entity.ReviewPointStatus;
