--- daebbang-core/src/main/java/com/daebbang/daebbangcore/domain/user/dto/request/LoginRequest.java
@@ -1,9 +1,0 @@
-package com.daebbang.daebbangcore.domain.user.dto.request;
-
-public record LoginRequest(
-
-    String id,
-    String password
-) {
-
-}
